// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/DailyRecurrencePatternDTO.ts
// This file was generated by TypeWriter. See TypeWriter.Common.tst for reference.



    export interface IDailyRecurrencePatternDTO {
        type: DailyRecurrencePatternType;
        interval: number;
        dayOfMonth: number;
        month: number;
        daysOfWeek: DayOfWeek[];
        firstDayOfWeek: DayOfWeek;
        weekIndex: DailyRecurrencePatternWeekIndex;
        sysHolidayTypeIds: number[];
    }

    

    

    export interface IDailyRecurrenceRangeDTO {
        type: DailyRecurrenceRangeType;
        startDate?: Date;
        endDate?: Date;
        numberOfOccurrences: number;
    }

    

    

    export interface IDailyRecurrenceDatesOutput {
        recurrenceDates: Date[];
        removedDates: Date[];
    }

    

    






import { DayOfWeek } from "../generated-interfaces/ClientEnumerations";
import { DailyRecurrencePatternType, DailyRecurrencePatternWeekIndex, DailyRecurrenceRangeType } from "../generated-interfaces/Enumerations";