// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/Enumerations.ts
export enum DailyRecurrencePatternType { 
    None = 0,
    Daily = 1,
    Weekly = 2,
    AbsoluteMonthly = 3,
    RelativeMonthly = 4,
    AbsoluteYearly = 5,
    RelativeYearly = 6,
    SysHoliday = 7
}

export enum DailyRecurrencePatternWeekIndex { 
    First = 0,
    Second = 1,
    Third = 2,
    Fourth = 3,
    Last = 4
}

export enum DailyRecurrenceRangeType { 
    EndDate = 0,
    NoEnd = 1,
    Numbered = 2
}