// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/ClientEnumerations.ts
// System
export enum DayOfWeek {
  Sunday = 0,
  Monday = 1,
  Tuesday = 2,
  Wednesday = 3,
  Thursday = 4,
  Friday = 5,
  Saturday = 6,
}

// SOEMessageBox
export enum SOEMessageBoxButtons {
  None = 0,
  OK = 1,
  OKCancel = 2,
  YesNo = 3,
  YesNoCancel = 4,
}

export enum SOEMessageBoxImage {
  None = 0,
  Information = 1,
  Warning = 2,
  Error = 3,
  Question = 4,
  Forbidden = 5,
  OK = 6,
  Spinner = 7,
  Custom = 8,
}